api/requirements.txt
aioredis==1.3.1
asgiref==3.4.1
async-timeout==4.0.1
attrs==21.2.0
autobahn==21.2.1
Automat==20.2.0
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
channels-redis==3.3.1
chardet==4.0.0
constantly==15.1.0
cryptography==36.0.0
daphne==3.0.2
dj-database-url==0.5.0
Django==3.2.9
django-cors-headers==3.10.0
django-db-logger==0.1.11
django-filter==21.1
django-ipware==4.0.0
django-rest-knox==4.1.0
django-user-agents==0.4.0
djangorestframework==3.12.4
freezegun==1.1.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.8.2
incremental==21.3.0
iniconfig==1.1.1
more-itertools==8.12.0
msgpack==1.0.3
packaging==21.3
pluggy==1.0.0
psycopg2-binary==2.9.2
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyHamcrest==2.0.2
pyOpenSSL==21.0.0
pyparsing==3.0.6
pytest==6.2.5
python-dateutil==2.8.2
python-decouple==3.5
pytz==2021.3
PyYAML==6.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
Twisted==21.7.0
txaio==21.2.1
typing-extensions==4.0.1
ua-parser==0.10.0
urllib3==1.26.7
user-agents==2.2.0
wcwidth==0.2.5
zipp==3.6.0
zope.interface==5.4.0
